requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
Django==4.1.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
Pillow==9.3.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.6
pyttsx3==2.90
pytweening==1.0.4
pywhatkit==5.4
requests==2.28.1
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
Werkzeug==2.2.2
wikipedia==1.4.0
